refactor: extract component accessor
